feat: add feedback button and modal
